Prepare HandleVehicleSpeedometer()

+++ TombEngine/Game/Hud/Speedometer.h
@@ -12,7 +12,7 @@ namespace TEN::Hud
 		bool _hasValueUpdated = false;
 
 		float _value		= 0.0f;
-		float _pointerAngle = 0.0f;
+		short _pointerAngle = 0;
 		float _opacity		= 0.0f;
 		float _life			= 0.0f;
